requirements.txt
asgiref==3.3.1
boto3==1.17.17
botocore==1.20.17
dj-database-url==0.5.0
Django==3.1.6
django-allauth==0.44.0
django-countries==7.0
django-crispy-forms==1.11.0
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
sqlparse==0.4.1
stripe==2.55.2
